requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing-extensions==4.9.0
uritemplate==4.1.1
